# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vilo AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CheckoutInput {
  userName: string;
  userEmail: string;
  /** @nullable */
  userPhone?: string | null;
}

export interface CheckoutResult {
  url: string;
}

export type ProvisioningStatusStatus = typeof ProvisioningStatusStatus[keyof typeof ProvisioningStatusStatus];


export const ProvisioningStatusStatus = {
  pending: 'pending',
  active: 'active',
  failed: 'failed',
} as const;

export interface ProvisioningStatus {
  status: ProvisioningStatusStatus;
  /** @nullable */
  viloNumber?: string | null;
}

export type AdminViloUserStatus = typeof AdminViloUserStatus[keyof typeof AdminViloUserStatus];


export const AdminViloUserStatus = {
  pending: 'pending',
  active: 'active',
  failed: 'failed',
} as const;

export interface AdminViloUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  viloNumber?: string | null;
  /** @nullable */
  twilioSid?: string | null;
  /** @nullable */
  elevenLabsPhoneId?: string | null;
  /** @nullable */
  stripeSessionId?: string | null;
  status: AdminViloUserStatus;
  createdAt: string;
}

export interface AdminUsersResult {
  users: AdminViloUser[];
  total: number;
}

export interface RetryResult {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}