# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vilo AI API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: checkout
    description: Stripe checkout operations
  - name: users
    description: User management
  - name: admin
    description: Admin operations (password protected)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /checkout:
    post:
      operationId: createCheckout
      tags: [checkout]
      summary: Create Stripe checkout session
      description: Creates a Stripe checkout session for Vilo AI subscription ($3.98)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/session/{sessionId}:
    get:
      operationId: getProvisioningStatus
      tags: [users]
      summary: Get provisioning status by Stripe session ID
      description: >
        Returns provisioning status and assigned Vilo number.
        Keyed on the Stripe checkout session ID (unguessable) rather than
        email, so no personal data is exposed to unauthenticated callers.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningStatus"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{id}/retry:
    post:
      operationId: retryProvisioning
      tags: [admin]
      summary: Retry provisioning for a failed user
      description: Resets a failed user back to pending and re-triggers Twilio + ElevenLabs provisioning. Requires Authorization header with admin password.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer <ADMIN_PASSWORD>"
      responses:
        "200":
          description: Retry started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetryResult"
        "400":
          description: User is not in failed status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all provisioned users
      description: Returns all Vilo users. Requires Authorization header with admin password.
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer <ADMIN_PASSWORD>"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CheckoutInput:
      type: object
      required: [userName, userEmail]
      properties:
        userName:
          type: string
        userEmail:
          type: string
        userPhone:
          type: ["string", "null"]

    CheckoutResult:
      type: object
      required: [url]
      properties:
        url:
          type: string

    ProvisioningStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, active, failed]
        viloNumber:
          type: ["string", "null"]

    AdminViloUser:
      type: object
      required: [id, name, email, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        viloNumber:
          type: ["string", "null"]
        twilioSid:
          type: ["string", "null"]
        elevenLabsPhoneId:
          type: ["string", "null"]
        stripeSessionId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, failed]
        createdAt:
          type: string

    AdminUsersResult:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminViloUser"
        total:
          type: integer

    RetryResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string